--- infra-sofa-boot-starter/src/main/java/com/alipay/boot/middleware/config/spring/namespace/spi/Slite2MiddlewareTagNameSupport.java
@@ -1,32 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alipay.boot.middleware.config.spring.namespace.spi;
-
-/**
- * 标签名
- *
- * Created by yangguanchao on 16/9/1.
- */
-public interface Slite2MiddlewareTagNameSupport {
-
-    /**
-     * 获取支持的 TagName
-     *
-     * @return 支持的 TagName
-     */
-    public String supportTagName();
-}
