+++ infra-sofa-boot-starter/src/main/java/com/alipay/sofa/infra/config/spring/namespace/spi/SofaBootTagNameSupport.java
@@ -16,13 +16,18 @@
  */
 package com.alipay.sofa.infra.config.spring.namespace.spi;
 
-import com.alipay.boot.middleware.config.spring.namespace.spi.Slite2MiddlewareTagNameSupport;
-
 /**
  * SofaBootTagNameSupport
  *
  * @author yangguanchao
  * @since 2018/04/05
  */
-public interface SofaBootTagNameSupport extends Slite2MiddlewareTagNameSupport {
+public interface SofaBootTagNameSupport {
+
+    /**
+     * 获取支持的 TagName
+     *
+     * @return 支持的 TagName
+     */
+    String supportTagName();
 }
