--- infra-sofa-boot-starter/src/main/resources/META-INF/com/alipay/boot/middleware/service/config/xts.xsd
@@ -1,111 +1,0 @@
-<!--
-  ~ Licensed to the Apache Software Foundation (ASF) under one or more
-  ~ contributor license agreements.  See the NOTICE file distributed with
-  ~ this work for additional information regarding copyright ownership.
-  ~ The ASF licenses this file to You under the Apache License, Version 2.0
-  ~ (the "License"); you may not use this file except in compliance with
-  ~ the License.  You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<xsd:schema xmlns="http://schema.alipay.com/sofa/schema/slite"
-            xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:beans="http://www.springframework.org/schema/beans"
-            targetNamespace="http://schema.alipay.com/sofa/schema/slite"
-            elementFormDefault="qualified" attributeFormDefault="unqualified">
-
-    <xsd:import namespace="http://www.springframework.org/schema/beans"/>
-
-    <!-- xts start -->
-    <xsd:element name="xts" type="Txts"/>
-    <xsd:complexType name="Txts">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:all minOccurs="1">
-                    <xsd:element name="datasource" type="TxtsDatasource" minOccurs="1" maxOccurs="1"/>
-                    <xsd:element name="dbtype" type="Tdbtype" minOccurs="0"/>
-                    <xsd:element name="activityTableName" type="TactivityTableName" minOccurs="0"/>
-                    <xsd:element name="actionTableName" type="TactionTableName" minOccurs="0"/>
-                    <xsd:element name="deleteMode" type="TdeleteMode" minOccurs="0"/>
-                    <xsd:element name="isDeleteDelayed" type="TisDeleteDelayed" minOccurs="0"/>
-                    <xsd:element name="idempotenceTableName" type="TidempotenceTableName" minOccurs="0"/>
-                    <xsd:element name="antiSuspendTableName" type="TantiSuspendTableName" minOccurs="0"/>
-                </xsd:all>
-                <xsd:attribute name="domain" type="xsd:string" use="optional"/>
-                <!--<xsd:attribute name="id" type="xsd:string" use="required"/>-->
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TxtsDatasource">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="ref" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="Tdbtype">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TactivityTableName">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TactionTableName">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TdeleteMode">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TisDeleteDelayed">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TidempotenceTableName">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TantiSuspendTableName">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <!-- xts end -->
-</xsd:schema>
