--- infra-sofa-boot-starter/src/main/resources/META-INF/com/alipay/boot/middleware/service/config/zdal.xsd
@@ -1,114 +1,0 @@
-<!--
-  ~ Licensed to the Apache Software Foundation (ASF) under one or more
-  ~ contributor license agreements.  See the NOTICE file distributed with
-  ~ this work for additional information regarding copyright ownership.
-  ~ The ASF licenses this file to You under the Apache License, Version 2.0
-  ~ (the "License"); you may not use this file except in compliance with
-  ~ the License.  You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<xsd:schema xmlns="http://schema.alipay.com/sofa/schema/slite"
-            xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:beans="http://www.springframework.org/schema/beans"
-            targetNamespace="http://schema.alipay.com/sofa/schema/slite"
-            elementFormDefault="qualified" attributeFormDefault="unqualified">
-
-    <xsd:import namespace="http://www.springframework.org/schema/beans"/>
-
-    <!-- zdal start -->
-    <xsd:element name="zdal-jdbc" type="Tzdal"/>
-    <xsd:complexType name="Tzdal">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:sequence minOccurs="1">
-                    <xsd:element name="appDataSourceName" type="TzdalDatasource" minOccurs="1" maxOccurs="1"/>
-                    <xsd:element name="configPath" type="TzdalDatasource" minOccurs="0" maxOccurs="1"/>
-                    <xsd:element name="elasticRuleBaseOnTable" type="TelasticRuleBaseOnTable" minOccurs="0"
-                                 maxOccurs="1"/>
-                    <xsd:element name="zdalConfigurationPath" type="TzdalConfigurationPath" minOccurs="0"
-                                 maxOccurs="1"/>
-                    <xsd:element name="sqlAdapter" type="TsqlAdapter" minOccurs="0" maxOccurs="1"/>
-                    <xsd:element name="translator" type="Ttranslator" minOccurs="0" maxOccurs="1"/>
-                    <xsd:element name="dataSourceFactory" type="TdataSourceFactory" minOccurs="0" maxOccurs="unbounded"/>
-                    <xsd:element name="dataSourceFactories" type="TdataSourceFactories" minOccurs="0" maxOccurs="1"/>
-                    <xsd:element name="isHintReplaceSupport" type="TisHintReplaceSupport" minOccurs="0" maxOccurs="1"/>
-                </xsd:sequence>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TzdalDatasource">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TelasticRuleBaseOnTable">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TzdalConfigurationPath">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TsqlAdapter">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="ref" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="Ttranslator">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="ref" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TdataSourceFactory">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="ref" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TdataSourceFactories">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:sequence minOccurs="1">
-                    <xsd:element name="dataSourceFactory" type="TdataSourceFactory" minOccurs="1" maxOccurs="unbounded"/>
-                </xsd:sequence>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-
-    <xsd:complexType name="TisHintReplaceSupport">
-        <xsd:complexContent>
-            <xsd:extension base="beans:identifiedType">
-                <xsd:attribute name="value" type="xsd:string" use="required"/>
-            </xsd:extension>
-        </xsd:complexContent>
-    </xsd:complexType>
-    <!-- zdal start -->
-
-</xsd:schema>
